requirements.txt
python-dotenv
imap-tools
pymupdf
psycopg2-binary
SQLAlchemy
alembic
caldav
icalendar